backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
pytesseract==0.3.10
PyMuPDF==1.23.8
google-generativeai==0.3.1
pydantic==2.5.0
pdf2image==1.16.3
aiofiles==23.2.1
python-magic==0.4.27